Thunk move to Redux from actions added name for user in DB models & migrations
