--- client/src/redux/actions/thunk/thunkCity.js
@@ -1,8 +1,0 @@
-import { getCity } from "../cityAction"
-import axios from 'axios'
-
-export const getCityFromDB = () =>  async (dispatch) => {
-  const response = await axios.get("http://localhost:5001/api/test")
-  console.log(response);
-  dispatch(getCity(response.data))
-}
