+++ client/src/redux/thunk/thunkCity.js
@@ -1,0 +1,8 @@
+import {getCity} from "../actions/cityAction"
+import axios from 'axios'
+
+export const getCityFromDB = () => async (dispatch) => {
+    const response = await axios.get("http://localhost:5001/api/test")
+    console.log(response);
+    dispatch(getCity(response.data))
+}
