+++ client/src/redux/thunk/thunkRegistration.js
@@ -1,5 +1,5 @@
-import AuthService from "../../../services/AuthService";
-import {getUser} from "../userActions";
+import AuthService from "../../services/AuthService";
+import {getUser} from "../actions/userActions";
 
 export const THUNK_ACTION_REGISTER = (userData) => async (dispatch) => {
     try {
