+++ server/db/models/user.js
@@ -1,30 +1,32 @@
 const {
-  Model,
+    Model,
 } = require('sequelize');
 
 module.exports = (sequelize, DataTypes) => {
-  class User extends Model {
-    /**
-     * Helper method for defining associations.
-     * This method is not a part of Sequelize lifecycle.
-     * The `models/index` file will call this method automatically.
-     */
-    static associate(models) {
-      // define association here
-      this.hasOne(models.Token, {
-        foreignKey: 'id',
+    class User extends Model {
+        /**
+         * Helper method for defining associations.
+         * This method is not a part of Sequelize lifecycle.
+         * The `models/index` file will call this method automatically.
+         */
+        static associate(models) {
+            // define association here
+            this.hasOne(models.Token, {
+                foreignKey: 'id',
 
-      });
+            });
+        }
     }
-  }
-  User.init({
-    email: DataTypes.STRING,
-    password: DataTypes.STRING,
-    isActivated: DataTypes.BOOLEAN,
-    activationLink: DataTypes.STRING,
-  }, {
-    sequelize,
-    modelName: 'User',
-  });
-  return User;
+
+    User.init({
+        name: DataTypes.STRING,
+        email: DataTypes.STRING,
+        password: DataTypes.STRING,
+        isActivated: DataTypes.BOOLEAN,
+        activationLink: DataTypes.STRING,
+    }, {
+        sequelize,
+        modelName: 'User',
+    });
+    return User;
 };
